task 103/ services get and update user skills, + hand
