--- sapphire/users/api/v1beta/rest/users/skkills/__init__.py
@@ -1,1 +1,0 @@
-from .router import router
