--- sapphire/users/api/v1beta/rest/users/skkills/router.py
@@ -1,7 +1,0 @@
-import fastapi
-
-from . import handlers
-
-router = fastapi.APIRouter()
-
-router.add_api_route(path="/skills", methods=["POST"],  endpoint=handlers.slills_installation)
